examples/bitcoin-price-tracker-0.7.1-snowflake/requirements.txt
# Python dependencies for Bitcoin Price Tracker - Snowflake version
# Install with: pip install -r requirements.txt

# Snowflake connector
snowflake-connector-python>=4.4.0

# API requests
requests>=2.32.5

# YAML parsing
pyyaml>=6.0

# Optional: pandas for bulk loading
pandas>=1.5.0

# Optional: FLUID CLI for declarative deployment
# fluid-forge>=0.7.1
